requirements.txt
aiohttp==3.9.1
asyncpg==0.29.0
Pillow==10.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
aiohttp-apispec==2.2.3
marshmallow==3.20.1